move ctx into wiggle-tests crate
